requirements.txt
pandas==1.0.1
seaborn==0.10.0
pillow
numpy>=1.18.1
matplotlib>=3.2.0
scipy>=1.4.1